exactbooks-backend/requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.0.12
dj-database-url==1.0.0
Django==3.2.4
django-bunny-storage==0.1.2
django-cockroachdb==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.14.0
drf-extra-fields==3.4.1
idna==3.4
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.5
requests==2.27.1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.12